Change double to float

+++ src/pl/sda/Main.java
@@ -4,10 +4,10 @@ public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner (System.in);
         System.out.println("Height in m:");
-        Double height = scanner.nextDouble();
+        float height = scanner.nextFloat();
         System.out.println("Wieght in kg:");
-        Double wieght = scanner.nextDouble();
-        Double BMI = wieght / (Math.pow(height,2));
+        float wieght = scanner.nextFloat();
+        double BMI = wieght / (Math.pow(height,2));
         System.out.println("Your BMI:" + BMI);
